+++ migration_odbc_import/doc/changelog.rst
@@ -1,3 +1,9 @@
+v1.1
+====
+* Simplify creation of connetion strings
+* Add primative table validation
+* View distinct values of a column (designed to assist importing data into selection fields)
+
 v1.0
 ====
-Initial Release
+* Initial Release